+++ src/snkmt/core/db/alembic/env.py
@@ -2,7 +2,7 @@ from logging.config import fileConfig
 
 from sqlalchemy import engine_from_config
 from sqlalchemy import pool
-from snkmt.db.models.base import Base
+from snkmt.core.models.base import Base
 from alembic import context
 
 # this is the Alembic Config object, which provides
