+++ src/snkmt/core/db/alembic/versions/c59016d243cc_.py
@@ -1,0 +1,32 @@
+"""empty message
+
+Revision ID: c59016d243cc
+Revises: a088a7b93fe5
+Create Date: 2025-08-30 13:26:32.962472
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'c59016d243cc'
+down_revision: Union[str, None] = 'a088a7b93fe5'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('snkmt_db_version', sa.Column('timestamp', sa.DateTime(), nullable=False))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('snkmt_db_version', 'timestamp')
+    # ### end Alembic commands ###
