+++ src/snkmt/core/models/version.py
@@ -1,15 +1,11 @@
 from typing import Any
-from snkmt.db.models.base import Base
+from snkmt.core.models.base import Base
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy import String
 from datetime import datetime, timezone
 
 # db versioning adapted from https://github.com/insitro/redun/blob/main/redun/backends/db/__init__.py
 DB_UNKNOWN_VERSION = 99
-
-
-class DBVersionError(Exception):
-    pass
 
 
 class DBVersion(Base):
@@ -58,33 +54,3 @@ class DBVersion(Base):
             return f"{self.major}.?"
         else:
             return f"{self.major}.{self.minor}"
-
-
-# List of all available database versions and migrations.
-# Note these are sorted from oldest to newest.
-null_db_version = DBVersion(id="000000000000", major=-1, minor=0)
-DB_VERSIONS = [DBVersion(id="a088a7b93fe5", major=1, minor=0)]
-DB_MIN_VERSION = DBVersion(id="", major=1, minor=0)  # Min db version needed by snkmt.
-DB_MAX_VERSION = DBVersion(id="", major=1, minor=0)  # Max db version needed by snkmt.
-
-
-def parse_db_version(version_str: str) -> DBVersion:
-    """
-    Parses a db version string such as "2.0" or "3" into a DBVersion.
-    """
-    if version_str == "latest":
-        return DB_VERSIONS[-1]
-
-    dots = version_str.count(".")
-    if dots == 0:
-        major, minor = (int(version_str), 0)
-    elif dots == 1:
-        major, minor = tuple(map(int, version_str.split(".")))
-    else:
-        raise ValueError(f"Invalid db version format: {version_str}")
-
-    for version in DB_VERSIONS:
-        if version.major == major and version.minor == minor:
-            return version
-
-    raise DBVersionError(f"Unknown db version: {version_str}")
