--- src/snkmt/db/models/__init__.py
@@ -1,19 +1,0 @@
-from snkmt.db.models.enums import Status, FileType
-
-
-from snkmt.db.models.workflow import Workflow
-from snkmt.db.models.rule import Rule
-from snkmt.db.models.job import Job
-from snkmt.db.models.file import File
-from snkmt.db.models.error import Error
-
-
-__all__ = [
-    "Status",
-    "FileType",
-    "Workflow",
-    "Rule",
-    "Job",
-    "File",
-    "Error",
-]
