+++ src/snkmt/types/enums.py
